+++ feature-teacher/register/src/commonMain/kotlin/com/b1nd/dodam/register/state/InfoEvent.kt
@@ -1,0 +1,7 @@
+package com.b1nd.dodam.register.state
+sealed interface InfoEvent {
+    data object SuccessGetAuthPhoneCode : InfoEvent
+    data object SuccessVerifyAuthPhoneCode : InfoEvent
+    data object FiledVerifyAuthCode : InfoEvent
+    data object FiledGetAuthCode : InfoEvent
+}
